+++ NeuralNetworksAndDeepLearning/ICost.cs
@@ -6,9 +6,9 @@ namespace NeuralNetworksAndDeepLearning
 {
     public interface ICost
     {
-        double Cost(double[] activations, double[] outputs);
-        double Function(double activation, double output);
-        double DelCostOverDelActivation(double activation, double output);
-        double DelCostOverDelWeightedInput(double weightedInput, double activation, double output);
+        float Cost(float[] activations, float[] outputs);
+        float Function(float activation, float output);
+        float DelCostOverDelActivation(float activation, float output);
+        float DelCostOverDelWeightedInput(float weightedInput, float activation, float output);
     }
 }
